Revert "Merge pull request #23575 Let Kotlin DSL standalone script templates use implicit receivers"
This reverts commit f56bc19f6ea2c430ab44b038658c4c110ffd2079, reversing
changes made to 9a14bd5cd426bdaf22fb5ff6f0ea51649599e135.
