+++ subprojects/architecture-test/src/changes/archunit_store/public-api-methods-return-allowed-types.txt
@@ -44,6 +44,9 @@ Method <org.gradle.kotlin.dsl.GroovyInteroperabilityKt.delegateClosureOf(java.la
 Method <org.gradle.kotlin.dsl.KotlinClosure0.getFunction()> has arguments/return type kotlin.jvm.functions.Function0 that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (GroovyInteroperability.kt:66)
 Method <org.gradle.kotlin.dsl.KotlinClosure2.getFunction()> has arguments/return type kotlin.jvm.functions.Function2 that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (GroovyInteroperability.kt:111)
 Method <org.gradle.kotlin.dsl.KotlinClosure3.getFunction()> has arguments/return type kotlin.jvm.functions.Function3 that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (GroovyInteroperability.kt:135)
+Method <org.gradle.kotlin.dsl.KotlinSettingsScriptKt.fileOperationsFor(org.gradle.api.initialization.Settings)> has arguments/return type org.gradle.api.internal.file.FileOperations that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (KotlinSettingsScript.kt:456)
+Method <org.gradle.kotlin.dsl.KotlinSettingsScriptKt.fileOperationsFor(org.gradle.api.invocation.Gradle, java.io.File)> has arguments/return type org.gradle.api.internal.file.FileOperations that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (KotlinSettingsScript.kt:461)
+Method <org.gradle.kotlin.dsl.KotlinSettingsScriptKt.fileOperationsFor(org.gradle.internal.service.ServiceRegistry, java.io.File)> has arguments/return type org.gradle.api.internal.file.FileOperations, org.gradle.internal.service.ServiceRegistry that are not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (KotlinSettingsScript.kt:466)
 Method <org.gradle.kotlin.dsl.precompile.PrecompiledScriptDependenciesResolver.resolve(kotlin.script.dependencies.ScriptContents, java.util.Map, kotlin.jvm.functions.Function3, kotlin.script.dependencies.KotlinScriptExternalDependencies)> has arguments/return type kotlin.script.dependencies.KotlinScriptExternalDependencies, kotlin.script.dependencies.ScriptContents, kotlin.jvm.functions.Function3 that are not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (PrecompiledScriptDependenciesResolver.kt:78)
 Method <org.gradle.language.assembler.tasks.Assemble.getOperationLoggerFactory()> has arguments/return type org.gradle.internal.operations.logging.BuildOperationLoggerFactory that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (Assemble.java:85)
 Method <org.gradle.language.base.sources.BaseLanguageSourceSet.create(java.lang.Class, java.lang.Class, org.gradle.platform.base.internal.ComponentSpecIdentifier, org.gradle.api.model.ObjectFactory)> has arguments/return type org.gradle.platform.base.internal.ComponentSpecIdentifier that is not Gradle public API or primitive or built-in JDK classes or Groovy classes or Kotlin classes in (BaseLanguageSourceSet.java:48)
