+++ subprojects/architecture-test/src/changes/archunit_store/public-api-not-extends-internal-types.txt
@@ -27,13 +27,11 @@ Class <org.gradle.kotlin.dsl.ArtifactHandlerScope> extends/implements org.gradle
 Class <org.gradle.kotlin.dsl.ClientModuleScope> extends/implements org.gradle.kotlin.dsl.support.delegates.ClientModuleDelegate that is Gradle Internal API in (DependencyHandlerExtensions.kt:0)
 Class <org.gradle.kotlin.dsl.DependencyConstraintHandlerScope> extends/implements org.gradle.kotlin.dsl.support.delegates.DependencyConstraintHandlerDelegate that is Gradle Internal API in (DependencyConstraintHandlerScope.kt:0)
 Class <org.gradle.kotlin.dsl.DependencyHandlerScope> extends/implements org.gradle.kotlin.dsl.support.delegates.DependencyHandlerDelegate that is Gradle Internal API in (DependencyHandlerScope.kt:0)
-Class <org.gradle.kotlin.dsl.InitScriptApi> extends/implements org.gradle.kotlin.dsl.support.delegates.GradleDelegate that is Gradle Internal API in (InitScriptApi.kt:0)
-Class <org.gradle.kotlin.dsl.KotlinBuildScript> extends/implements org.gradle.kotlin.dsl.support.DefaultKotlinScript that is Gradle Internal API in (KotlinBuildScript.kt:0)
-Class <org.gradle.kotlin.dsl.KotlinInitScript> extends/implements org.gradle.kotlin.dsl.support.DefaultKotlinScript that is Gradle Internal API in (KotlinInitScript.kt:0)
-Class <org.gradle.kotlin.dsl.KotlinSettingsScript> extends/implements org.gradle.kotlin.dsl.support.DefaultKotlinScript that is Gradle Internal API in (KotlinSettingsScript.kt:0)
+Class <org.gradle.kotlin.dsl.InitScriptApi> extends/implements org.gradle.kotlin.dsl.support.delegates.GradleDelegate that is Gradle Internal API in (KotlinInitScript.kt:0)
+Class <org.gradle.kotlin.dsl.KotlinBuildScript> extends/implements org.gradle.kotlin.dsl.support.delegates.ProjectDelegate that is Gradle Internal API in (KotlinBuildScript.kt:0)
 Class <org.gradle.kotlin.dsl.NamedDomainObjectContainerScope> extends/implements org.gradle.kotlin.dsl.support.delegates.NamedDomainObjectContainerDelegate that is Gradle Internal API in (NamedDomainObjectContainerExtensions.kt:0)
 Class <org.gradle.kotlin.dsl.ScriptHandlerScope> extends/implements org.gradle.kotlin.dsl.support.delegates.ScriptHandlerDelegate that is Gradle Internal API in (ScriptHandlerScope.kt:0)
-Class <org.gradle.kotlin.dsl.SettingsScriptApi> extends/implements org.gradle.kotlin.dsl.support.delegates.SettingsDelegate that is Gradle Internal API in (SettingsScriptApi.kt:0)
+Class <org.gradle.kotlin.dsl.SettingsScriptApi> extends/implements org.gradle.kotlin.dsl.support.delegates.SettingsDelegate that is Gradle Internal API in (KotlinSettingsScript.kt:0)
 Class <org.gradle.kotlin.dsl.TaskContainerScope> extends/implements org.gradle.kotlin.dsl.support.delegates.TaskContainerDelegate that is Gradle Internal API in (TaskContainerExtensions.kt:0)
 Class <org.gradle.kotlin.dsl.precompile.PrecompiledProjectScript> extends/implements org.gradle.kotlin.dsl.support.delegates.ProjectDelegate that is Gradle Internal API in (PrecompiledProjectScript.kt:0)
 Class <org.gradle.language.base.sources.BaseLanguageSourceSet> extends/implements org.gradle.language.base.internal.AbstractLanguageSourceSet that is Gradle Internal API in (BaseLanguageSourceSet.java:0)
