+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/KotlinBuildScript.kt
@@ -17,46 +17,54 @@
 package org.gradle.kotlin.dsl
 
 import org.gradle.api.Project
-import org.gradle.plugin.use.PluginDependenciesSpec
+import org.gradle.api.initialization.dsl.ScriptHandler
+
 import org.gradle.kotlin.dsl.resolver.KotlinBuildScriptDependenciesResolver
-import org.gradle.kotlin.dsl.support.DefaultKotlinScript
 import org.gradle.kotlin.dsl.support.KotlinScriptHost
-import org.gradle.kotlin.dsl.support.defaultKotlinScriptHostForProject
+import org.gradle.kotlin.dsl.support.delegates.ProjectDelegate
 import org.gradle.kotlin.dsl.support.internalError
 import org.gradle.kotlin.dsl.support.invalidPluginsCall
 import org.gradle.kotlin.dsl.template.KotlinBuildScriptTemplateAdditionalCompilerArgumentsProvider
-import kotlin.script.experimental.annotations.KotlinScript
-import kotlin.script.experimental.api.baseClass
-import kotlin.script.experimental.api.filePathPattern
-import kotlin.script.experimental.api.implicitReceivers
+
+import org.gradle.plugin.use.PluginDependenciesSpec
+
+import kotlin.script.extensions.SamWithReceiverAnnotations
 import kotlin.script.templates.ScriptTemplateAdditionalCompilerArguments
 import kotlin.script.templates.ScriptTemplateDefinition
 
 
-private
-class KotlinBuildScriptCompilationConfiguration : KotlinDslStandaloneScriptCompilationConfiguration({
-    filePathPattern.put(".+(?<!(^|\\.)(init|settings))\\.gradle\\.kts")
-    baseClass(KotlinBuildScript::class)
-    implicitReceivers(Project::class)
-})
-
-
 /**
- * Base class for Gradle Kotlin DSL standalone [Project] scripts, aka. build scripts.
+ * Base class for Kotlin build scripts.
  */
-@KotlinScript(
-    compilationConfiguration = KotlinBuildScriptCompilationConfiguration::class
-)
 @ScriptTemplateDefinition(
     resolver = KotlinBuildScriptDependenciesResolver::class,
+    scriptFilePattern = ".+(?<!(^|\\.)(init|settings))\\.gradle\\.kts"
 )
 @ScriptTemplateAdditionalCompilerArguments(
+    [
+        "-language-version", "1.8",
+        "-api-version", "1.8",
+        "-Xjvm-default=all",
+        "-Xjsr305=strict",
+        "-XXLanguage:+DisableCompatibilityModeForNewInference",
+        "-XXLanguage:-TypeEnhancementImprovementsInStrictMode",
+    ],
     provider = KotlinBuildScriptTemplateAdditionalCompilerArgumentsProvider::class
 )
+@SamWithReceiverAnnotations("org.gradle.api.HasImplicitReceiver")
 @GradleDsl
 abstract class KotlinBuildScript(
-    host: KotlinScriptHost<Project>
-) : DefaultKotlinScript(defaultKotlinScriptHostForProject(host.target)) {
+    private val host: KotlinScriptHost<Project>
+) : ProjectDelegate() /* TODO:kotlin-dsl configure Project as implicit receiver */ {
+
+    override val delegate: Project
+        get() = host.target
+
+    /**
+     * The [ScriptHandler] for this script.
+     */
+    override fun getBuildscript(): ScriptHandler =
+        host.scriptHandler
 
     /**
      * Configures the build script classpath for this project.
@@ -64,7 +72,7 @@ abstract class KotlinBuildScript(
      * @see [Project.buildscript]
      */
     @Suppress("unused")
-    open fun buildscript(block: ScriptHandlerScope.() -> Unit): Unit =
+    open fun buildscript(@Suppress("unused_parameter") block: ScriptHandlerScope.() -> Unit): Unit =
         internalError()
 
     /**
