+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/precompile/v1/PrecompiledScriptTemplates.kt
@@ -18,16 +18,20 @@ package org.gradle.kotlin.dsl.precompile.v1
 
 import org.gradle.api.Project
 import org.gradle.api.initialization.Settings
+import org.gradle.api.internal.ProcessOperations
+import org.gradle.api.internal.file.FileOperations
 import org.gradle.api.invocation.Gradle
+import org.gradle.api.logging.Logger
+import org.gradle.api.logging.Logging
+import org.gradle.api.logging.LoggingManager
 import org.gradle.api.plugins.PluginAware
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ProviderConvertible
 import org.gradle.kotlin.dsl.*
 import org.gradle.kotlin.dsl.precompile.PrecompiledScriptDependenciesResolver
 import org.gradle.kotlin.dsl.support.DefaultKotlinScript
-import org.gradle.kotlin.dsl.support.defaultKotlinScriptHostForGradle
 import org.gradle.kotlin.dsl.support.defaultKotlinScriptHostForProject
-import org.gradle.kotlin.dsl.support.defaultKotlinScriptHostForSettings
+import org.gradle.kotlin.dsl.support.serviceOf
 import org.gradle.plugin.use.PluginDependenciesSpec
 import org.gradle.plugin.use.PluginDependency
 import org.gradle.plugin.use.PluginDependencySpec
@@ -63,7 +67,16 @@ import kotlin.script.templates.ScriptTemplateDefinition
 @GradleDsl
 open class PrecompiledInitScript(
     target: Gradle
-) : DefaultKotlinScript(defaultKotlinScriptHostForGradle(target)), PluginAware by target
+) : DefaultKotlinScript(InitScriptHost(target)), PluginAware by target {
+
+    private
+    class InitScriptHost(val gradle: Gradle) : Host {
+        override fun getLogger(): Logger = Logging.getLogger(Gradle::class.java)
+        override fun getLogging(): LoggingManager = gradle.serviceOf()
+        override fun getFileOperations(): FileOperations = fileOperationsFor(gradle, null)
+        override fun getProcessOperations(): ProcessOperations = gradle.serviceOf()
+    }
+}
 
 
 /**
@@ -80,7 +93,7 @@ open class PrecompiledInitScript(
 @GradleDsl
 open class PrecompiledSettingsScript(
     target: Settings
-) : DefaultKotlinScript(defaultKotlinScriptHostForSettings(target)), PluginAware by target {
+) : DefaultKotlinScript(SettingsScriptHost(target)), PluginAware by target {
 
     /**
      * Configures the plugin dependencies for this settings script.
@@ -106,6 +119,14 @@ open class PrecompiledSettingsScript(
                 }
             }
         )
+    }
+
+    private
+    class SettingsScriptHost(val settings: Settings) : Host {
+        override fun getLogger(): Logger = Logging.getLogger(Settings::class.java)
+        override fun getLogging(): LoggingManager = settings.serviceOf()
+        override fun getFileOperations(): FileOperations = fileOperationsFor(settings)
+        override fun getProcessOperations(): ProcessOperations = settings.serviceOf()
     }
 }
