+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/DefaultKotlinScript.kt
@@ -25,13 +25,10 @@ import org.gradle.api.file.ConfigurableFileTree
 import org.gradle.api.file.CopySpec
 import org.gradle.api.file.DeleteSpec
 import org.gradle.api.file.FileTree
-import org.gradle.api.initialization.Settings
 import org.gradle.api.internal.ProcessOperations
 import org.gradle.api.internal.file.FileOperations
 import org.gradle.api.internal.project.ProjectInternal
-import org.gradle.api.invocation.Gradle
 import org.gradle.api.logging.Logger
-import org.gradle.api.logging.Logging
 import org.gradle.api.logging.LoggingManager
 import org.gradle.api.resources.ResourceHandler
 import org.gradle.api.tasks.WorkResult
@@ -90,7 +87,6 @@ open class DefaultKotlinScript internal constructor(
                 @Suppress("unchecked_cast")
                 fileOperations.fileTree(baseDir as Map<String, *>)
             }
-
             else -> fileOperations.fileTree(baseDir)
         }
 
@@ -147,32 +143,4 @@ class ProjectScriptHost(val project: Project) : DefaultKotlinScript.Host {
     override fun getFileOperations(): FileOperations = projectInternal().fileOperations
     override fun getProcessOperations(): ProcessOperations = projectInternal().processOperations
     fun projectInternal() = (project as ProjectInternal)
-}
-
-
-internal
-fun defaultKotlinScriptHostForSettings(settings: Settings): DefaultKotlinScript.Host =
-    SettingsScriptHost(settings)
-
-
-private
-class SettingsScriptHost(val settings: Settings) : DefaultKotlinScript.Host {
-    override fun getLogger(): Logger = Logging.getLogger(Settings::class.java)
-    override fun getLogging(): LoggingManager = settings.serviceOf()
-    override fun getFileOperations(): FileOperations = fileOperationsFor(settings)
-    override fun getProcessOperations(): ProcessOperations = settings.serviceOf()
-}
-
-
-internal
-fun defaultKotlinScriptHostForGradle(gradle: Gradle): DefaultKotlinScript.Host =
-    GradleScriptHost(gradle)
-
-
-private
-class GradleScriptHost(val gradle: Gradle) : DefaultKotlinScript.Host {
-    override fun getLogger(): Logger = Logging.getLogger(Gradle::class.java)
-    override fun getLogging(): LoggingManager = gradle.serviceOf()
-    override fun getFileOperations(): FileOperations = fileOperationsFor(gradle, null)
-    override fun getProcessOperations(): ProcessOperations = gradle.serviceOf()
 }
