+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/ArtifactHandlerDelegate.kt
@@ -26,6 +26,8 @@ import org.gradle.api.artifacts.dsl.ArtifactHandler
 
 /**
  * Facilitates the implementation of the [ArtifactHandler] interface by delegation via subclassing.
+ *
+ * See [GradleDelegate] for why this is currently necessary.
  */
 abstract class ArtifactHandlerDelegate : ArtifactHandler {
