+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/ClientModuleDelegate.kt
@@ -37,6 +37,8 @@ import org.gradle.api.capabilities.Capability
 
 /**
  * Facilitates the implementation of the [ClientModule] interface by delegation via subclassing.
+ *
+ * See [GradleDelegate] for why this is currently necessary.
  */
 abstract class ClientModuleDelegate : ClientModule {
