+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/ScriptHandlerDelegate.kt
@@ -31,6 +31,8 @@ import java.net.URI
 
 /**
  * Facilitates the implementation of the [ScriptHandler] interface by delegation via subclassing.
+ *
+ * See [GradleDelegate] for why this is currently necessary.
  */
 abstract class ScriptHandlerDelegate : ScriptHandler {
