+++ subprojects/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/delegates/SettingsDelegate.kt
@@ -41,7 +41,6 @@ import java.io.File
 /**
  * Facilitates the implementation of the [Settings] interface by delegation via subclassing.
  */
-@Deprecated("Will be removed in Gradle 9.0")
 abstract class SettingsDelegate : Settings {
 
     internal
